+++ dlls/crowbar.cpp
@@ -183,7 +183,7 @@ int CCrowbar::Swing( int fFirst )
 #endif
 	PLAYBACK_EVENT_FULL( FEV_NOTHOST, m_pPlayer->edict(), m_usCrowbar, 
 	0.0, (float *)&g_vecZero, (float *)&g_vecZero, 0, 0, 0,
-	0.0, 0, 0.0 );
+	0, 0, 0 );
 
 	if( tr.flFraction >= 1.0 )
 	{
@@ -294,7 +294,7 @@ int CCrowbar::Swing( int fFirst )
 			m_trHit = tr;
 		}
 
-		m_pPlayer->m_iWeaponVolume = flVol * CROWBAR_WALLHIT_VOLUME;
+		m_pPlayer->m_iWeaponVolume = (int)( flVol * CROWBAR_WALLHIT_VOLUME );
 #endif
 		m_flNextPrimaryAttack = GetNextAttackDelay( 0.25 );
