+++ dlls/effects.h
@@ -238,7 +238,7 @@ public:
 
 	inline int GetWidth( void )
 	{
-		return pev->scale;
+		return (int)pev->scale;
 	}
 
 	inline int GetNoise( void )
@@ -255,17 +255,17 @@ public:
 
 	inline int GetBrightness( void )
 	{
-		return pev->renderamt;
+		return (int)pev->renderamt;
 	}
 
 	inline int GetFrame( void )
 	{
-		return pev->frame;
+		return (int)pev->frame;
 	}
 
 	inline int GetScrollRate( void )
 	{
-		return pev->animtime;
+		return (int)pev->animtime;
 	}
 
 	// Call after you change start/end positions
